Refactored page service to create shared component
