+++ client/src/components/game/lab5/pages/NotificationActivity/NotificationAccessibleRepairKnowledgeCheck.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import KnowledgeTest from "../../components/KnowledgeTest";
-import PageServiceTimer from "../../components/PageServiceTimer";
+import PageServiceTimer from "../../../shared/PageServiceTimer";
 import {NotificationAccessibleRepairKnowledgeCheck as KnowledgeCheck} from '../../../../../constants/lab5'
 class NotificationAccessibleRepairKnowledgeCheck extends Component {
