+++ client/src/components/game/lab5/pages/PageLayoutActivity/DyslexiaAccessibleKnowledgeCheck.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import KnowledgeTest from "../../components/KnowledgeTest";
-import PageServiceTimer from "../../components/PageServiceTimer";
+import PageServiceTimer from "../../../shared/PageServiceTimer";
 import {DyslexiaAccessibleKnowledgeCheck as KnowledgeCheck} from '../../../../../constants/lab5'
 class DyslexiaAccessibleKnowledgeCheck extends Component {
